ajuste de bugs

+++ Projeto de Cartas/cartas.c
@@ -5,7 +5,7 @@ int main() {
     char estado1;
     char codigo1[5];
     char cidade1[50];
-    int populacao1;
+    float populacao1;
     float area1;
     float pib1;
     int pontos1;
@@ -14,7 +14,7 @@ int main() {
     char estado2;
     char codigo2[5];
     char cidade2[50];
-    int populacao2;
+    float populacao2;
     float area2; 
     float pib2;
     int pontos2;
@@ -34,7 +34,7 @@ int main() {
     scanf(" %[^\n]", cidade1);
 
     printf("População:\n> ");
-    scanf("%d", &populacao1);
+    scanf("%f", &populacao1);
 
     printf("Área (em km²): \n> ");
     scanf("%f", &area1);
@@ -58,7 +58,7 @@ int main() {
     scanf(" %[^\n]", cidade2);
 
     printf("População:\n> ");
-    scanf("%d", &populacao2);
+    scanf("%f", &populacao2);
 
     printf("Área (em km²): \n> ");
     scanf("%f", &area2);
@@ -77,18 +77,18 @@ int main() {
     printf("Estado: %c\n", estado1);
     printf("Código: %s\n", codigo1);
     printf("Cidade: %s\n", cidade1);
-    printf("População: %d\n", populacao1);
-    printf("Área: %.2f km²\n", area1);
-    printf("PIB: %.2f bilhões\n", pib1);
+    printf("População: %.3f\n", populacao1);
+    printf("Área: %.3f km²\n", area1);
+    printf("PIB: %.3f bilhões\n", pib1);
     printf("Pontos Turísticos: %d\n", pontos1);
 
     printf("\nCarta 2:\n");
     printf("Estado: %c\n", estado2);
     printf("Código: %s\n", codigo2);
     printf("Cidade: %s\n", cidade2);
-    printf("População: %d\n", populacao2);
-    printf("Área: %.2f km²\n", area2);
-    printf("PIB: %.2f bilhões\n", pib2);
+    printf("População: %.3f\n", populacao2);
+    printf("Área: %.3f km²\n", area2);
+    printf("PIB: %.3f bilhões\n", pib2);
     printf("Pontos Turísticos: %d\n", pontos2);
 
     return 0;
